Enhancement: Extract IsEmptyTest
