+++ tests/Constraint/ConstraintTest.php
@@ -1744,57 +1744,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintIsEmpty()
-    {
-        $constraint = new IsEmpty;
-
-        $this->assertFalse($constraint->evaluate(['foo'], '', true));
-        $this->assertTrue($constraint->evaluate([], '', true));
-        $this->assertFalse($constraint->evaluate(new \ArrayObject(['foo']), '', true));
-        $this->assertTrue($constraint->evaluate(new \ArrayObject([]), '', true));
-        $this->assertEquals('is empty', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate(['foo']);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that an array is empty.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintIsEmpty2()
-    {
-        $constraint = new IsEmpty;
-
-        try {
-            $constraint->evaluate(['foo'], 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message\nFailed asserting that an array is empty.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintCountWithAnArray()
     {
         $constraint = new Count(5);
