+++ tests/Constraint/IsEmptyTest.php
@@ -1,0 +1,69 @@
+<?php
+/*
+ * This file is part of sebastian/phpunit-framework-constraint.
+ *
+ * (c) Sebastian Bergmann <sebastian@phpunit.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PHPUnit\Framework\Constraint;
+
+use PHPUnit\Framework\ExpectationFailedException;
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\TestFailure;
+
+class IsEmptyTest extends TestCase
+{
+    public function testConstraintIsEmpty()
+    {
+        $constraint = new IsEmpty();
+
+        $this->assertFalse($constraint->evaluate(['foo'], '', true));
+        $this->assertTrue($constraint->evaluate([], '', true));
+        $this->assertFalse($constraint->evaluate(new \ArrayObject(['foo']), '', true));
+        $this->assertTrue($constraint->evaluate(new \ArrayObject([]), '', true));
+        $this->assertEquals('is empty', $constraint->toString());
+        $this->assertCount(1, $constraint);
+
+        try {
+            $constraint->evaluate(['foo']);
+        } catch (ExpectationFailedException $e) {
+            $this->assertEquals(
+                <<<EOF
+Failed asserting that an array is empty.
+
+EOF
+                ,
+                TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    public function testConstraintIsEmpty2()
+    {
+        $constraint = new IsEmpty();
+
+        try {
+            $constraint->evaluate(['foo'], 'custom message');
+        } catch (ExpectationFailedException $e) {
+            $this->assertEquals(
+                <<<EOF
+custom message\nFailed asserting that an array is empty.
+
+EOF
+                ,
+                TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+}
